+++ R/tsibble-package.R
@@ -80,9 +80,6 @@
 #' @importFrom lubridate quarter is.POSIXct rollback force_tz tz is.POSIXt is.Date
 #' @importFrom lubridate decimal_date date_decimal round_date
 #' @importFrom lubridate seconds_to_period month<- weeks days as_date
-#' @importFrom purrr map map_dbl map_int map_chr map_lgl reduce imap
-#' @importFrom purrr map2 map2_dbl map2_int map2_chr map2_lgl
-#' @importFrom purrr pmap pmap_dbl pmap_int pmap_chr pmap_lgl
 #' @importFrom dplyr arrange filter select mutate transmute summarise rename
 #' @importFrom dplyr group_by ungroup group_data grouped_df group_vars
 #' @importFrom dplyr group_rows groups new_grouped_df is_grouped_df