+++ tests/testthat/test-yearweek.R
@@ -121,34 +121,27 @@ test_that("yearweek() with missing `by` #228", {
 })
 
 test_that("yearweek() set operations", {
-  expect_identical(intersect(yearweek("2020 W1") + 0:6,
-                             yearweek("2020 W1") + 3:9),
-                   yearweek("2020 W1") + 3:6)
+  expect_identical(
+    intersect(yearweek("2020 W1") + 0:6, yearweek("2020 W1") + 3:9),
+    yearweek("2020 W1") + intersect(0:6, 3:9))
 
-  expect_error(intersect(yearweek("2020 W1") + 0:6, Sys.Date()),
-               "'y' must be of class 'yearweek'")
   expect_error(intersect(
     yearweek("2020 W1"),
     yearweek("2020 W1", week_start = 2)
   ))
 
-  expect_identical(union(yearweek("2020 W1") + 0:6,
-                         yearweek("2020 W1") + 3:9),
-                   yearweek("2020 W1") + 0:9)
+  expect_identical(
+    union(yearweek("2020 W1") + 0:6, yearweek("2020 W1") + 3:9),
+    yearweek("2020 W1") + 0:9)
 
-  expect_error(union(yearweek("2020 W1") + 0:6, Sys.Date()),
-               "'y' must be of class 'yearweek'")
   expect_error(union(
     yearweek("2020 W1"),
     yearweek("2020 W1", week_start = 2)
   ))
 
-  expect_identical(setdiff(yearweek("2020 W1") + 0:6,
-                           yearweek("2020 W1") + 3:9),
-                   yearweek("2020 W1") + c(0:2, 7:9))
-
-  expect_error(setdiff(yearweek("2020 W1") + 0:6, Sys.Date()),
-               "'y' must be of class 'yearweek'")
+  expect_identical(
+    setdiff(yearweek("2020 W1") + 0:6, yearweek("2020 W1") + 3:9),
+    yearweek("2020 W1") + setdiff(0:6, 3:9))
   expect_error(setdiff(
     yearweek("2020 W1"),
     yearweek("2020 W1", week_start = 2)