+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fixed C stack error when index contains NA with `as_tsibble()`. (#229)
 * Fixed bug when `by` is missing in `year*()`. (#228)
 * Removed rolling functions.
+* Removed dependency of {purrr}.
 
 # tsibble 0.9.3